requirements/scrapers.txt
BeautifulSoup==3.2.1
python-dateutil==2.2
requests==2.0.1
